app/requirements.txt
flask 
pyzbar 
pdf2image 
cfenv 
gunicorn 
PyPDF2 
requests_toolbelt 
sap_cf_logging
sap-xssec